-CodeGen improvements -Assembler additions -Emit options and release mode for code gen -Additions and changes to the codefile format -Adding hex numbers in Print[ln]Va functions

+++ jkr/CodeFile/Function.h
@@ -11,11 +11,12 @@ enum FunctionAttributes {
 constexpr auto MaxArguments = 32;
 
 struct FunctionHeader {
-    Byte Attributes;
+    Uint16 Attributes;
     // Arguments are part of LocalCount but are used for diferent things
+    // If you use the reg callconv you
     Byte Arguments;
     Byte LocalCount;
-    unsigned int SizeOfCode;
+    Uint16 SizeOfCode;
 };
 
 }
